Add option for multiple tags
Given that we handle tags differently between production and
development, the publish of tags has to be more flexible than just the
`latest`.

To tackle that, I've added a `-t` option that'll receive multiple tags if
specified more than once. However, as you might have noticed, there's a
clash with the current option for git versioning. Therefore, we renamed
the git option from `-t` to `-v` with no value; thus, you can think that if
there's a `v` on _options_, the script will **v**ersion your code.

Refers-to: tagging, versioning, docker