Separate versioning and publishing
As the current process consists of publishing a duplicated image, with
the same version tag, to production and development; the arrangement of
scripts obstructs it, since the version is created and published in one
shot, without a chance for reuse.

To address that, now it's a developer's responsibility to create the
version with `create-version` and later pass the returned value to a tag
with `publish ... -t "$version"`. Developers needed to remember that
before when specifying `-v` to publish, but now versioning and
publishing are detached. Also, to make the transition easier and
increase the adoption, I've added an example for `bitbucket-pipelines`
with the process used throughout the company.

Refers-to: tagging, versioning, examples, publishing