Allow publishing of multiple images on a single call
Depending on how the repository and images correlate, especially on
monorepos, developers should be capable of publishing multiple images
using the same project, tags, etc.

We changed the parameter `-i` from a single string to an array of
strings, thus enabling the use of multiple `-i` parameters. Also, as
we already had a mechanism to load a file named `image.docker` by
default to help with multiple steps on bitbucket, I've extended that
functionality by loading any file named `$image.docker`, following the
image names which the caller wants to publish.